+++ docs/en/docs/release-notes.md
@@ -11,6 +11,7 @@ hide:
 
 - Mark `http_exception` in the `Exception` as `async` to avoid thread creation.
 - Replace `asyncio.iscoroutinefunction` with `inspect.iscoroutinefunction`.
+- Replace `asyncio` in `run_sync` with anyio.
 
 ## 0.16.10